py2/django/rest/requirements.txt
markdown

# django & plugins
django
djangorestframework
django-filter
django-extensions
